re-enable joint noise adding
re-enable joint noise adding
