+++ src/hermes.cpp
@@ -1,5 +1,4 @@
 #include "emp-sh2pc/emp-sh2pc.h"
-#include <utils/row.h>
 #include <utils/io.h>
 #include <utils/oblisort.h>
 #include <utils/memory.h>
@@ -24,63 +23,6 @@ using std::chrono::duration_cast;
 using std::chrono::duration;
 using std::chrono::milliseconds;
 
-
-
-// Configurations
-int col_length0 = 64; 
-
-
-void tcpds_q1_standard(Data *left, Data *right, uint32_t key_pos, uint32_t contr){
-	auto t1 = high_resolution_clock::now();
-	for (u_int32_t i=0; i < TPCDS_MAXSTEPS; i++){
-		//std::cout << "Testing round -> "<< i <<std::endl;
-		op_join(left, right, 0, 1);
-		std::cout <<"Current round ->" <<i<<std::endl;
-	}
-	auto t2 = high_resolution_clock::now();
-	/* Getting number of milliseconds as a double. */
-    duration<double, std::milli> ms_double = t2 - t1;
-	std::cout << "Total time for tpcds-q1:" << ms_double.count() << "ms" <<endl;
-}
-
-
-
-void test_millionare(int party, int number) {
-	Integer a(32, number, ALICE);
-	Integer b(32, number, BOB);
-	Bit res = a > b;
-
-	cout << "ALICE larger?\t"<< res.reveal<bool>()<<endl;
-}
-
-void test_sort(int party) {
-	int size = 100;
-	Integer *A = new Integer[size];
-	Integer *B = new Integer[size];
-	Integer *res = new Integer[size];
-
-// First specify Alice's input
-	for(int i = 0; i < size; ++i)
-		A[i] = Integer(32, rand()%102400, ALICE);
-
-
-// Now specify Bob's input
-	for(int i = 0; i < size; ++i)
-		B[i] = Integer(32, rand()%102400, BOB);
-
-//Now compute
-	for(int i = 0; i < size; ++i)
-		res[i] = A[i] ^ B[i];
-	
-
-	sort(res, size);
-	for(int i = 0; i < 100; ++i)
-		cout << res[i].reveal<int32_t>()<<endl;
-
-	delete[] A;
-	delete[] B;
-	delete[] res;
-}
 
 int main(int argc, char** argv) {
 	int port, party;
@@ -108,23 +50,6 @@ int main(int argc, char** argv) {
 
 	op_filter_gt(data, prev_cnt, 5, party);
 	io->flush();
-
-#ifdef	DEBUG_JOIN
-	std::vector<int32_t> vect;
-	vect = fnfetch_data(full_path + "0.txt");
-	std::vector<int32_t> vect2;
-	vect2 = fnfetch_data(full_path + "1.txt");
-
-	Data *data = op_recover(vect, attr, 0, party);
-	io->flush();
-
-	Data *data2 = op_recover(vect2, attr, 0, party);
-	io->flush();
-
-	//op_merg_join(data, data2, 0, 1);
-	//op_sort(data2->data, data2->public_size, Bit(false));
-	io->flush();
-#endif
 
 	//cout << CircuitExecution::circ_exec->num_and()<<endl;
 	finalize_semi_honest();
