+++ utils/data.h
@@ -1,0 +1,15 @@
+#ifndef DATA_H__
+#define DATA_H__
+
+using namespace emp;
+using namespace std;
+// define maximum size of values array
+#define MAX_VALUES 5
+
+// define struct for data
+struct data {
+    Integer flag;
+    Integer values[MAX_VALUES];
+};
+
+#endif
